テストコードのリファクタリング、勝敗結果を記録する処理
